Move into parser package for library usage
