+++ parser/parser.go.y
@@ -1,5 +1,5 @@
 %{
-package main
+package parser
 
 %}
